% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_r0.R
\name{calc_r0}
\alias{calc_r0}
\title{Calculate R0 for Norovirus}
\usage{
calc_r0(t1, t2, data)
}
\arguments{
\item{t1}{The starting epidemiological week (as a Date)}

\item{t2}{The ending epidemiological week (as a Date)}

\item{data}{A data frame with 'epi_week' and 'count' columns}
}
\value{
The calculated R0 value
}
\description{
Calculate R0 for Norovirus
}
