% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_change_points_cusum.R
\name{detect_change_points_cusum}
\alias{detect_change_points_cusum}
\title{Detect Change Points using CUSUM and Plot Results}
\usage{
detect_change_points_cusum(data, threshold = 5)
}
\arguments{
\item{data}{A data frame containing a time series of data with `epi_week` and `count` columns.}

\item{threshold}{The CUSUM threshold value, initially set to a reasonable value, defaults to 5.}
}
\value{
A ggplot object with change points marked, and prints a summary of the CUSUM analysis.
}
\description{
Detect Change Points using CUSUM and Plot Results
}
