% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_change_points_growth_rate.R
\name{detect_change_points_growth_rate}
\alias{detect_change_points_growth_rate}
\title{Detect Change Points using Growth Rate Analysis and Plot Results}
\usage{
detect_change_points_growth_rate(data, threshold = 0.5)
}
\arguments{
\item{data}{A data frame containing a time series of data with `epi_week` and `count` columns.}

\item{threshold}{The growth rate threshold for detecting change points, defaults to 0.5 (50% increase).}
}
\value{
A ggplot object with change points marked, and prints the dates of change points.
}
\description{
Detect Change Points using Growth Rate Analysis and Plot Results
}
