% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_change_points_pelt.R
\name{detect_change_points_pelt}
\alias{detect_change_points_pelt}
\title{Detect Change Points using PELT and Plot Results}
\usage{
detect_change_points_pelt(data, penalty = "BIC")
}
\arguments{
\item{data}{A data frame containing a time series of data with `epi_week` and `count` columns.}

\item{penalty}{The penalty value used in the PELT method to avoid overfitting, defaults to "BIC".}
}
\value{
A ggplot object with change points marked, and prints the dates of change points.
}
\description{
Detect Change Points using PELT and Plot Results
}
