% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_growth_rate_changes.R
\name{detect_growth_rate_changes}
\alias{detect_growth_rate_changes}
\title{Detect Sharp Increases in Growth Rate and Plot Results}
\usage{
detect_growth_rate_changes(data, threshold = 10)
}
\arguments{
\item{data}{A data frame containing a time series of data with `epi_week` and `count` columns.}

\item{threshold}{The percentage growth rate threshold to identify sharp increases, defaults to 10 (10% increase).}
}
\value{
A ggplot object with change points marked, and prints a summary of the growth rate analysis.
}
\description{
Detect Sharp Increases in Growth Rate and Plot Results
}
