% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_onset_exponential_rise.R
\name{identify_onset_exponential_rise}
\alias{identify_onset_exponential_rise}
\title{Identify the Onset of Exponential Rises in Seasonal Peaks}
\usage{
identify_onset_exponential_rise(data, nsmo = 9, threshold_multiplier = 1)
}
\arguments{
\item{data}{A data frame containing a time series with `epi_week` and `count` columns.}

\item{nsmo}{The window size for the moving average to smooth the data, defaults to 9.}

\item{threshold_multiplier}{Multiplier for setting the threshold based on the standard deviation, defaults to 1.}
}
\value{
A ggplot object showing the time series with onsets marked.
}
\description{
Identify the Onset of Exponential Rises in Seasonal Peaks
}
