% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_onset_exponential_rise_1.R
\name{identify_onset_exponential_rise_1}
\alias{identify_onset_exponential_rise_1}
\title{Identify the Onset of Exponential Rises in Seasonal Peaks with Smoothing}
\usage{
identify_onset_exponential_rise_1(data, nsmo = 9, threshold = NULL)
}
\arguments{
\item{data}{A data frame containing a time series with `epi_week` and `count` columns.}

\item{nsmo}{The window size for the moving average to smooth the data, defaults to 9.}

\item{threshold}{The threshold for the rate of change to detect an onset, automatically set based on the smoothed data.}
}
\value{
A ggplot object showing the time series with onsets marked.
}
\description{
Identify the Onset of Exponential Rises in Seasonal Peaks with Smoothing
}
