% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_troughs.R
\name{identify_troughs}
\alias{identify_troughs}
\title{Identify the Onset of Exponential Rises by Derivative Analysis}
\usage{
identify_troughs(data, nsmo = 9)
}
\arguments{
\item{data}{A data frame containing a time series with `epi_week` and `count` columns.}

\item{nsmo}{The window size for the moving average to smooth the data, defaults to 9.}
}
\value{
A ggplot object showing the time series with onsets marked.
}
\description{
Identify the Onset of Exponential Rises by Derivative Analysis
}
