% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cases_w_boxcar.R
\name{plot_cases_w_boxcar}
\alias{plot_cases_w_boxcar}
\title{Plot Norovirus Cases with a Running Average}
\usage{
plot_cases_w_boxcar(data, nsmo = 9, ytit = "Number of Cases")
}
\arguments{
\item{data}{A data frame containing the norovirus data, with `epi_week` and `count` columns.}

\item{nsmo}{The number of weeks to use for the running average, defaults to 9.}

\item{ytit}{The y-axis title for the plot, defaults to "Number of Cases".}
}
\value{
A ggplot object
}
\description{
Plot Norovirus Cases with a Running Average
}
