% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power_spectrum.R
\name{plot_power_spectrum}
\alias{plot_power_spectrum}
\title{Plot Power Spectrum with Annotated Peaks}
\usage{
plot_power_spectrum(data, npeak = 2)
}
\arguments{
\item{data}{A data frame containing the norovirus data with a `count` column.}

\item{npeak}{The number of peaks to annotate, defaults to 2.}
}
\value{
The ggplot object showing the power spectrum plot.
}
\description{
Plot Power Spectrum with Annotated Peaks
}
