% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sarima_forecast.R
\name{plot_sarima_forecast}
\alias{plot_sarima_forecast}
\title{Plot SARIMA Forecast with Confidence Intervals}
\usage{
plot_sarima_forecast(data, forecast_length = 12)
}
\arguments{
\item{data}{A data frame containing the norovirus data with a `count` column and `epi_week` (Date type).}

\item{forecast_length}{The number of future data points to forecast (default is 12, for 12 weeks).}
}
\value{
The ggplot object showing the forecast plot.
}
\description{
Plot SARIMA Forecast with Confidence Intervals
}
