% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_norovirus_data.R
\name{read_norovirus_data}
\alias{read_norovirus_data}
\title{Read Norovirus Data from Excel File}
\usage{
read_norovirus_data(main_sheet_index = 2)
}
\arguments{
\item{main_sheet_index}{The index of the sheet to read (default is 2)}
}
\value{
A data frame with norovirus data
}
\description{
Read Norovirus Data from Excel File
}
